# src/GuildSaber.Website/openapi/GuildSaber.Api.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "GuildSaber.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/debug/import-old-gs-maps/{guildId}": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "Import ranked maps from old GuildSaber",
        "description": "Import ranked maps from old GuildSaber for a specific guild.",
        "operationId": "ImportOldGuildSaberMapsBackground",
        "parameters": [
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "example": {
                  "type": "https://tools.ietf.org/html/rfc9110#section-15.5.5",
                  "title": "Not Found",
                  "status": 404,
                  "detail": null,
                  "instance": "POST debug/import-old-gs-maps/{guildId}",
                  "traceId": "00-0123456789abcdef0123456789abcdef-0123456789abcdef-00"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
    "/debug/import-beatleader-scores/{playerId}": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "Enqueue BeatLeader player scores import.",
        "description": "Queues a background task to import all scores for the specified player from BeatLeader.",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
    "/debug/import-scoresaber-scores/{playerId}": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "Enqueue ScoreSaber player scores import.",
        "description": "Queues a background task to import all scores for the specified player from ScoreSaber.",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
    "/debug/remove-all-scores": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "Remove all scores from the database.",
        "description": "Deletes all ranked scores from the database. USE WITH CAUTION!",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
    "/debug/remove-all-ranked-scores": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "Remove all ranked scores from the database.",
        "description": "Deletes all ranked scores from the database. USE WITH CAUTION!",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
    "/debug/remove-all-ranked-maps": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "Remove all ranked maps from the database.",
        "description": "Deletes all ranked maps from the database. USE WITH CAUTION!",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
    "/debug/recalculate-member-points/{playerId}": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "Recalculate member points for a player.",
        "description": "Recalculates member points for all contexts the player is a member of.",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
    "/debug/recalculate-member-achievements/{playerId}": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "Recalculate member achievements for a player.",
        "description": "Recalculates member achievements for all contexts the player is a member of.",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
    "/debug/recalculate-player-scores/{playerId}": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "Recalculate player scores.",
        "description": "Recalculates all player scores for the specified player.",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
    "/debug/recalculate-all-player-scores": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "Recalculate all player scores.",
        "description": "Recalculates all player scores for all players in the database. USE WITH CAUTION!",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
    "/debug/refetch-all-player-scores-from-bl-and-scoresaber": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "Refetch all player scores from BeatLeader and ScoreSaber.",
        "description": "Refetches all player scores for all players in the database from BeatLeader and ScoreSaber. USE WITH CAUTION!",
        "parameters": [
          {
            "name": "fromPlayerId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
    "/debug/import-all-admin-conf": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "Import all admin confirmations from the old GuildSaber system.",
        "description": "Imports all admin confirmations from the old GuildSaber system for all players and guilds. USE WITH CAUTION!",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
    "/debug/delete-member-point-stats/{playerId}": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "Delete all member point stats for a player.",
        "description": "Deletes all member point stats for the specified player. USE WITH CAUTION!",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
    "/debug/trigger-edit-ranked-map-pipeline/{rankedMapId}": {
      "post": {
        "tags": [
          "Debug"
        ],
        "summary": "Trigger the EditRankedMapPipeline for a specific ranked map.",
        "description": "Triggers the EditRankedMapPipeline for the specified ranked map",
        "parameters": [
          {
            "name": "rankedMapId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "SessionCookie": [ ],
            "X-GuildSaber-Request": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IResult": {
        "type": "object"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "traceId": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "description": "GuildSaber session cookie. Browsers and cookie-enabled HTTP clients send it automatically.",
        "name": "__Host-GuildSaber.Session",
        "in": "cookie"
      },
      "X-GuildSaber-Request": {
        "type": "apiKey",
        "description": "Required for state-changing requests authenticated with the session cookie.",
        "name": "X-GuildSaber-Request",
        "in": "header"
      }
    }
  },
...
}